+++ test/e2e/chaos/basic.go
@@ -25,7 +25,6 @@ import (
 	"time"
 
 	"github.com/onsi/ginkgo"
-	"github.com/pingcap/errors"
 	corev1 "k8s.io/api/core/v1"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -49,6 +48,7 @@ import (
 	"github.com/chaos-mesh/chaos-mesh/test/e2e/util/portforward"
 	"github.com/chaos-mesh/chaos-mesh/test/pkg/fixture"
 
+	iochaostestcases "github.com/chaos-mesh/chaos-mesh/test/e2e/chaos/iochaos"
 	// testcases
 	podchaostestcases "github.com/chaos-mesh/chaos-mesh/test/e2e/chaos/podchaos"
 	timechaostestcases "github.com/chaos-mesh/chaos-mesh/test/e2e/chaos/timechaos"
@@ -188,175 +188,11 @@ var _ = ginkgo.Describe("[Basic]", func() {
 		ginkgo.Context("[IODelay]", func() {
 
 			ginkgo.It("[Schedule]", func() {
-				ctx, cancel := context.WithCancel(context.Background())
-				err = util.WaitE2EHelperReady(c, port)
-				framework.ExpectNoError(err, "wait e2e helper ready error")
-
-				ioChaos := &v1alpha1.IoChaos{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "io-chaos",
-						Namespace: ns,
-					},
-					Spec: v1alpha1.IoChaosSpec{
-						Selector: v1alpha1.SelectorSpec{
-							Namespaces:     []string{ns},
-							LabelSelectors: map[string]string{"app": "io"},
-						},
-						Action:     v1alpha1.IoLatency,
-						Mode:       v1alpha1.OnePodMode,
-						VolumePath: "/var/run/data",
-						Path:       "/var/run/data/*",
-						Delay:      "1s",
-						Percent:    100,
-						Duration:   pointer.StringPtr("9m"),
-						Scheduler: &v1alpha1.SchedulerSpec{
-							Cron: "@every 10m",
-						},
-					},
-				}
-				err = cli.Create(ctx, ioChaos)
-				framework.ExpectNoError(err, "create io chaos error")
-
-				err = wait.PollImmediate(5*time.Second, 1*time.Minute, func() (bool, error) {
-					dur, _ := getPodIODelay(c, port)
-
-					second := dur.Seconds()
-					klog.Infof("get io delay %fs", second)
-					// IO Delay >= 1s
-					if second >= 1 {
-						return true, nil
-					}
-					return false, nil
-				})
-				framework.ExpectNoError(err, "io chaos doesn't work as expected")
-				klog.Infof("apply io chaos successfully")
-
-				err = cli.Delete(ctx, ioChaos)
-				framework.ExpectNoError(err, "failed to delete io chaos")
-
-				err = wait.PollImmediate(5*time.Second, 1*time.Minute, func() (bool, error) {
-					dur, _ := getPodIODelay(c, port)
-
-					second := dur.Seconds()
-					klog.Infof("get io delay %fs", second)
-					// IO Delay shouldn't longer than 1s
-					if second >= 1 {
-						return false, nil
-					}
-					return true, nil
-				})
-				framework.ExpectNoError(err, "fail to recover io chaos")
-				cancel()
+				iochaostestcases.TestcaseIODelayDurationForATimeThenRecover(ns, cli, c, port)
 			})
 
 			ginkgo.It("[Pause]", func() {
-				ctx, cancel := context.WithCancel(context.Background())
-				err = util.WaitE2EHelperReady(c, port)
-				framework.ExpectNoError(err, "wait e2e helper ready error")
-
-				ioChaos := &v1alpha1.IoChaos{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "io-chaos",
-						Namespace: ns,
-					},
-					Spec: v1alpha1.IoChaosSpec{
-						Selector: v1alpha1.SelectorSpec{
-							Namespaces:     []string{ns},
-							LabelSelectors: map[string]string{"app": "io"},
-						},
-						Action:     v1alpha1.IoLatency,
-						Mode:       v1alpha1.OnePodMode,
-						VolumePath: "/var/run/data",
-						Path:       "/var/run/data/*",
-						Delay:      "1s",
-						Percent:    100,
-						Duration:   pointer.StringPtr("9m"),
-						Scheduler: &v1alpha1.SchedulerSpec{
-							Cron: "@every 10m",
-						},
-					},
-				}
-				err = cli.Create(ctx, ioChaos)
-				framework.ExpectNoError(err, "error occurs while applying io chaos")
-
-				err = wait.PollImmediate(5*time.Second, 1*time.Minute, func() (bool, error) {
-					dur, _ := getPodIODelay(c, port)
-
-					second := dur.Seconds()
-					klog.Infof("get io delay %fs", second)
-					// IO Delay >= 1s
-					if second >= 1 {
-						return true, nil
-					}
-					return false, nil
-				})
-				framework.ExpectNoError(err, "io chaos doesn't work as expected")
-
-				chaosKey := types.NamespacedName{
-					Namespace: ns,
-					Name:      "io-chaos",
-				}
-
-				// pause experiment
-				err = util.PauseChaos(ctx, cli, ioChaos)
-				framework.ExpectNoError(err, "pause chaos error")
-
-				err = wait.Poll(5*time.Second, 5*time.Minute, func() (done bool, err error) {
-					chaos := &v1alpha1.IoChaos{}
-					err = cli.Get(ctx, chaosKey, chaos)
-					framework.ExpectNoError(err, "get io chaos error")
-					if chaos.Status.Experiment.Phase == v1alpha1.ExperimentPhasePaused {
-						return true, nil
-					}
-					return false, err
-				})
-				framework.ExpectNoError(err, "check paused chaos failed")
-
-				// wait 1 min to check whether io delay still exists
-				err = wait.PollImmediate(5*time.Second, 1*time.Minute, func() (bool, error) {
-					dur, _ := getPodIODelay(c, port)
-
-					second := dur.Seconds()
-					klog.Infof("get io delay %fs", second)
-					// IO Delay shouldn't longer than 1s
-					if second >= 1 {
-						return false, nil
-					}
-					return true, nil
-				})
-				framework.ExpectNoError(err, "fail to recover io chaos")
-
-				// resume experiment
-				err = util.UnPauseChaos(ctx, cli, ioChaos)
-				framework.ExpectNoError(err, "resume chaos error")
-
-				err = wait.Poll(5*time.Second, 1*time.Minute, func() (done bool, err error) {
-					chaos := &v1alpha1.IoChaos{}
-					err = cli.Get(ctx, chaosKey, chaos)
-					framework.ExpectNoError(err, "get io chaos error")
-					if chaos.Status.Experiment.Phase == v1alpha1.ExperimentPhaseRunning {
-						return true, nil
-					}
-					return false, err
-				})
-				framework.ExpectNoError(err, "check resumed chaos failed")
-
-				err = wait.PollImmediate(5*time.Second, 1*time.Minute, func() (bool, error) {
-					dur, _ := getPodIODelay(c, port)
-
-					second := dur.Seconds()
-					klog.Infof("get io delay %fs", second)
-					// IO Delay >= 1s
-					if second >= 1 {
-						return true, nil
-					}
-					return false, nil
-				})
-				framework.ExpectNoError(err, "io chaos doesn't work as expected")
-
-				// cleanup
-				cli.Delete(ctx, ioChaos)
-				cancel()
+				iochaostestcases.TestcaseIODelayDurationForATimePauseAndUnPause(ns, cli, c, port)
 			})
 		})
 
@@ -364,171 +200,10 @@ var _ = ginkgo.Describe("[Basic]", func() {
 		ginkgo.Context("[IOErrno]", func() {
 
 			ginkgo.It("[Schedule]", func() {
-				ctx, cancel := context.WithCancel(context.Background())
-				err = util.WaitE2EHelperReady(c, port)
-				framework.ExpectNoError(err, "wait e2e helper ready error")
-
-				ioChaos := &v1alpha1.IoChaos{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "io-chaos",
-						Namespace: ns,
-					},
-					Spec: v1alpha1.IoChaosSpec{
-						Selector: v1alpha1.SelectorSpec{
-							Namespaces:     []string{ns},
-							LabelSelectors: map[string]string{"app": "io"},
-						},
-						Action:     v1alpha1.IoFaults,
-						Mode:       v1alpha1.OnePodMode,
-						VolumePath: "/var/run/data",
-						Path:       "/var/run/data/*",
-						Percent:    100,
-						// errno 5 is EIO -> I/O error
-						Errno: 5,
-						// only inject write method
-						Methods:  []v1alpha1.IoMethod{v1alpha1.Write},
-						Duration: pointer.StringPtr("9m"),
-						Scheduler: &v1alpha1.SchedulerSpec{
-							Cron: "@every 10m",
-						},
-					},
-				}
-				err = cli.Create(ctx, ioChaos)
-				framework.ExpectNoError(err, "create io chaos")
-
-				err = wait.PollImmediate(5*time.Second, 1*time.Minute, func() (bool, error) {
-					_, err = getPodIODelay(c, port)
-					// input/output error is errno 5
-					if err != nil && strings.Contains(err.Error(), "input/output error") {
-						return true, nil
-					}
-					return false, nil
-				})
-				framework.ExpectNoError(err, "io chaos doesn't work as expected")
-
-				err = cli.Delete(ctx, ioChaos)
-				framework.ExpectNoError(err, "failed to delete io chaos")
-
-				klog.Infof("success to perform io chaos")
-				err = wait.PollImmediate(5*time.Second, 1*time.Minute, func() (bool, error) {
-					_, err = getPodIODelay(c, port)
-
-					if err == nil {
-						return true, nil
-					}
-					return false, nil
-				})
-				framework.ExpectNoError(err, "fail to recover io chaos")
-
-				cancel()
+				iochaostestcases.TestcaseIOErrorDurationForATimeThenRecover(ns, cli, c, port)
 			})
-
 			ginkgo.It("[Pause]", func() {
-				ctx, cancel := context.WithCancel(context.Background())
-				err = util.WaitE2EHelperReady(c, port)
-				framework.ExpectNoError(err, "wait e2e helper ready error")
-
-				ioChaos := &v1alpha1.IoChaos{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "io-chaos",
-						Namespace: ns,
-					},
-					Spec: v1alpha1.IoChaosSpec{
-						Selector: v1alpha1.SelectorSpec{
-							Namespaces:     []string{ns},
-							LabelSelectors: map[string]string{"app": "io"},
-						},
-						Action:     v1alpha1.IoFaults,
-						Mode:       v1alpha1.OnePodMode,
-						VolumePath: "/var/run/data",
-						Path:       "/var/run/data/*",
-						Percent:    100,
-						// errno 5 is EIO -> I/O error
-						Errno: 5,
-						// only inject write method
-						Methods:  []v1alpha1.IoMethod{v1alpha1.Write},
-						Duration: pointer.StringPtr("9m"),
-						Scheduler: &v1alpha1.SchedulerSpec{
-							Cron: "@every 10m",
-						},
-					},
-				}
-				err = cli.Create(ctx, ioChaos)
-				framework.ExpectNoError(err, "create io chaos error")
-
-				klog.Info("create iochaos successfully")
-
-				err = wait.PollImmediate(5*time.Second, 1*time.Minute, func() (bool, error) {
-					_, err = getPodIODelay(c, port)
-					// input/output error is errno 5
-					if err != nil && strings.Contains(err.Error(), "input/output error") {
-						return true, nil
-					}
-					return false, nil
-				})
-				framework.ExpectNoError(err, "io chaos doesn't work as expected")
-
-				chaosKey := types.NamespacedName{
-					Namespace: ns,
-					Name:      "io-chaos",
-				}
-
-				// pause experiment
-				err = util.PauseChaos(ctx, cli, ioChaos)
-				framework.ExpectNoError(err, "pause chaos error")
-
-				klog.Info("pause iochaos")
-
-				err = wait.Poll(5*time.Second, 5*time.Minute, func() (done bool, err error) {
-					chaos := &v1alpha1.IoChaos{}
-					err = cli.Get(ctx, chaosKey, chaos)
-					framework.ExpectNoError(err, "get io chaos error")
-					if chaos.Status.Experiment.Phase == v1alpha1.ExperimentPhasePaused {
-						return true, nil
-					}
-					return false, err
-				})
-				framework.ExpectNoError(err, "check paused chaos failed")
-
-				// wait 1 min to check whether io delay still exists
-				err = wait.PollImmediate(5*time.Second, 1*time.Minute, func() (bool, error) {
-					_, err = getPodIODelay(c, port)
-
-					if err == nil {
-						return true, nil
-					}
-					return false, nil
-				})
-				framework.ExpectNoError(err, "fail to recover io chaos")
-
-				// resume experiment
-				err = util.UnPauseChaos(ctx, cli, ioChaos)
-				framework.ExpectNoError(err, "resume chaos error")
-
-				err = wait.Poll(5*time.Second, 1*time.Minute, func() (done bool, err error) {
-					chaos := &v1alpha1.IoChaos{}
-					err = cli.Get(ctx, chaosKey, chaos)
-					framework.ExpectNoError(err, "get io chaos error")
-					if chaos.Status.Experiment.Phase == v1alpha1.ExperimentPhaseRunning {
-						return true, nil
-					}
-					return false, err
-				})
-				framework.ExpectNoError(err, "check resumed chaos failed")
-
-				err = wait.PollImmediate(5*time.Second, 1*time.Minute, func() (bool, error) {
-					_, err = getPodIODelay(c, port)
-					// input/output error is errno 5
-					if err != nil && strings.Contains(err.Error(), "input/output error") {
-						return true, nil
-					}
-					return false, nil
-				})
-				framework.ExpectNoError(err, "io chaos doesn't work as expected")
-
-				// cleanup
-				cli.Delete(ctx, ioChaos)
-				cancel()
+				iochaostestcases.TestcaseIOErrorDurationForATimePauseAndUnPause(ns, cli, c, port)
 			})
 		})
 	})
@@ -1107,31 +782,6 @@ selector:
 	})
 
 })
-
-// get pod io delay
-func getPodIODelay(c http.Client, port uint16) (time.Duration, error) {
-	resp, err := c.Get(fmt.Sprintf("http://localhost:%d/io", port))
-	if err != nil {
-		return 0, err
-	}
-
-	out, err := ioutil.ReadAll(resp.Body)
-	defer resp.Body.Close()
-	if err != nil {
-		return 0, err
-	}
-
-	result := string(out)
-	if strings.Contains(result, "failed to write file") {
-		return 0, errors.New(result)
-	}
-	dur, err := time.ParseDuration(result)
-	if err != nil {
-		return 0, err
-	}
-
-	return dur, nil
-}
 
 func testNetworkDelay(c http.Client, port uint16, targetIP string) (int64, error) {
 	body := []byte(fmt.Sprintf("{\"targetIP\":\"%s\"}", targetIP))
